requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make requirements
#
-e git+git://github.com/dimagi/architect@v0.5.7a2#egg=architect
alembic==0.8.6
amqp==1.4.9
anyjson==0.3.3            # via kombu
appdirs==1.4.3            # via zeep
asn1crypto==0.24.0        # via cryptography
attrs==18.1.0
babel==2.6.0              # via django-phonenumber-field
billiard==3.3.0.23        # via celery
boto3==1.4.4
botocore==1.5.95          # via boto3, s3transfer
cached-property==1.4.3    # via zeep
celery==3.1.18
cffi==1.11.5              # via cryptography, csiphash
chardet==3.0.4            # via ghdiff
cloudant==2.7.0           # via jsonobject-couchdbkit
concurrentloghandler==0.9.1
contextlib2==0.5.4
cryptography==2.3         # via pyopenssl
csiphash==0.0.5
csv342==1.0.0
cython==0.27.3
datadog==0.15.0
decorator==4.0.11
defusedxml==0.4.1
diff-match-patch==20120106
dimagi-memoized==1.1.1
django-angular==0.7.16
django-appconf==1.0.2     # via django-compressor, django-statici18n
django-braces==1.8.1
django-bulk-update==1.1.10
django-celery==3.2.1
django-compressor==2.1
django-countries==4.6
django-crispy-forms==1.7.0
django-cte==1.1.4
django-formtools==2.0
django-otp==0.3.13
django-partial-index==0.4.0
django-phonenumber-field==1.3.0  # via django-two-factor-auth
django-prbac==0.0.7
django-ranged-response==0.2.0  # via django-simple-captcha
django-redis-sessions==0.5.6
django-redis==4.2
django-simple-captcha==0.5.7
django-statici18n==1.3.0
django-tastypie==0.14.0
django-transfer==0.4
django-two-factor-auth==1.6.2
django-user-agents==0.3.2
django-websocket-redis==0.5.1
django==1.11.14
djangorestframework==3.5.4
dnspython==1.15.0
docutils==0.14            # via botocore
dropbox==7.3.1
elasticsearch==1.9.0
email_validator==1.0.2
enum34==1.1.6             # via cryptography
et-xmlfile==1.0.1         # via openpyxl
ethiopian-date-converter==0.1.5
eulxml==1.1.3
faker==0.8.15
funcsigs==1.0.2           # via mock
futures==3.2.0            # via s3transfer, tinys3
gevent==1.2.2
ghdiff==0.4
greenlet==0.4.12
gunicorn==19.4
hiredis==0.2.0
html5lib==1.0b8
http-parser==0.8.3        # via restkit
httpagentparser==1.7.8
idna==2.7                 # via cryptography, email-validator
ipaddress==1.0.22         # via cryptography, faker
iso8601==0.1.11
isodate==0.6.0            # via zeep
jdcal==1.4                # via openpyxl
jinja2==2.8
jmespath==0.9.3           # via boto3, botocore
json-delta==2.0
jsonfield==1.0.3
jsonobject-couchdbkit==0.9.4
jsonobject==0.9.2
jsonpath-rw==1.4.0
kafka-python==0.9.5
kombu==3.0.37             # via celery
laboratory==0.2.0
lxml==3.4.4
mako==1.0.7               # via alembic
markdown==2.2.1
markupsafe==1.0           # via jinja2, mako
mock==2.0.0
msgpack-python==0.5.6     # via django-redis
ndg-httpsclient==0.5.1    # via requests
numpy==1.7.1
openpyxl==2.5.3
pbr==4.2.0                # via mock
phonenumberslite==8.8.9
pillow==5.2.0
ply==3.11                 # via eulxml, jsonpath-rw
polib==1.1.0
psycogreen==1.0
psycopg2==2.7.3.1
py-kissmetrics==1.0.1
pyasn1==0.4.3             # via ndg-httpsclient, requests
pycco==0.5.1
pycparser==2.18           # via cffi
pycrypto==2.6.1
pygithub==1.35
pygments==2.0.2
pygooglechart==0.4.0
pyjwt==1.6.4              # via pygithub, twilio
pyopenssl==18.0.0         # via ndg-httpsclient, requests
pypdf2==1.26.0            # via xhtml2pdf
pystache==0.5.4           # via pycco
python-dateutil==2.6.1
python-editor==1.0.3      # via alembic
python-magic==0.4.13
python-mimeparse==1.6.0   # via django-tastypie
pytz==2017.2
pyyaml==3.12
pyzxcvbn==0.8.0
qrcode==4.0.4
quickcache==0.5.1
raven==6.1.0
rcssmin==1.0.6            # via django-compressor
redis==2.10.5
reportlab==3.0
requests-toolbelt==0.8.0  # via zeep
requests[security]==2.10.0
restkit==4.2.2
rjsmin==1.0.12            # via django-compressor
s3transfer==0.1.13        # via boto3
sh==1.09
simpleeval==0.9.5
simplejson==3.11.1
six==1.11.0
smartypants==2.0.1        # via pycco
socketpool==0.5.3
sqlagg==0.9.1
sqlalchemy==1.2.9
stripe==1.67.0
suds-jurko==0.6
text-unidecode==1.2       # via faker
tinys3==0.1.12
tropo-webapi-python==0.1.3
twilio==6.5.1
ua-parser==0.8.0          # via user-agents
unidecode==0.04.20
urllib3==1.23             # via dropbox, elasticsearch, py-kissmetrics
user-agents==1.1.0        # via django-user-agents
werkzeug==0.11.15
xhtml2pdf==0.0.6
xlrd==1.0.0
xlwt==1.3.0
zeep==1.6.0
